https://codedrills.io/problems/the-married-unmarried-riddle
There's a quite famous riddle that goes like this:

Alice, Bob and Charles are standing in a line. Alice is looking at Bob. Bob is looking at Charles. We know that Alice is married, while Charles is not. Is there a married person among the three who is looking at an unmarried person?

Surprisingly, not many people get the riddle right. An unaware person would say that the answer to the riddle cannot be determined, as the marital status of Bob is unknown. However, look closely, and you'll find that the answer to the riddle is always 'Yes'. If Bob is married, he'll be the married person looking at unmarried Charles. If Bob is unmarried, Alice will be the married person looking at unmarried Bob.

This riddle was too easy for you, and you have decided to modify it:

N people are now standing in a line. For each valid i (1 ≤ i < N), the i'th person is looking at the (i + 1)'th person. Their marital statuses are given as a string S of length N, consisting of characters 'M' (married), 'U' (unmarried), '?' (unknown). The i'th character (1 ≤ i ≤ N) of string S corresponds to the i'th person standing in the line.

Given S, can you tell if there's a married person looking at an unmarried person, regardless of what the ?'s in S are replaced with?