https://codedrills.io/contests/icpc-amritapuri-2020-preliminary-round/problems/express-the-number
Given two integers n and x, we want to express n as a sum of k integers n = A[1] + A[2] + A[3] + ... + A[k] satisfying the following:

Either 0 ≤ A[1] ≤ x, or A[1] is an odd power of 2. The odd powers of 2 are 21, 23, 25, etc.
All the other A[i]s are odd powers of 2 for 2 ≤ i ≤ k.

Find the minimum possible value of k, or print -1 if we can't express n in the specified format.